Add config option to realtime read all tags
For some applications including demos it may be nessisary to always read all
tags from the m2web api instead of datamailbox.  To faciliate this in
a user friendly way that is always enabled a checkbox has been added to the
config page to force usage of the m2web api over datamailbox.