Added release fallback to unsigned modl
For customer specific features we want to be able to generate
intermediate releases quickly. For these we can use unsigned releases.
If the signed version of the module does not exist we fall back on the
unsigned.  Unsigned releases will be noted by the "-unsigned" at the
end of the filename.