Update Apache HTTP dependency
Replace the existing httpasyncclient dependency with
the newer httpclient5 dependency.

The httpclient5 dependency is more up-to-date and
makes use of Java 11 features. This will also allow
for a cleaner implementation of request classes.

Removed constant(s), class(es), and utility method(s)
that are now unused or rendered useless after the
dependency update.