Add M2WebEwonEBDResponse.java class
This class contains common JSON fields which
may be present in an Ewon EBD response if and
only if the M2Web API itself returns an error.

Each individual EBD response class has been
updated to try parsing the response for the
common M2Web error JSON fields if parsing as
CSV fails. Note: If the response cannot be
parsed as the applicable CSV or general JSON
format, an exception will still be thrown.